Ignore `dyn.NilValue` when traversing value from `dyn.Map`
The map function ignores cases where either a key in a map is not present or an
index in a sequence is out of bounds. As of recently, we retain nil values as
valid values in a configuration tree. As such, it makes sense to also ignore
cases where a map or sequence is expected but a nil is found. This is
semantically no different from an empty map where a key is not found.

Without this fix, all calls to `dyn.Map` would need to be updated with
nil-checks at every path component.