Fix glob expansion after running a generic build command
This didn't work as expected because the generic build mutator was calling into
the type specific build mutator in the middle of the function. This invalidates
the `config.Artifact` pointer that was being mutated later on, effectively
hiding these mutations from its caller.

To fix this, I turned glob expansion into its own mutator. It now works as
expected, _and_ produces better errors if the glob patterns are invalid or do
not match files.