Add paths field to bundle sync configuration
This field allows a user to configure paths to synchronize to the workspace.

Allowed values are relative paths to files and directories, anchored at the
directory where the field is set. If one or more values traverse up the
directory tree (to an ancestor of the bundle root directory), the CLI will
dynamically figure out the root path to use to ensure that the file tree
structure remains intact.

For example, given a `databricks.yml` in `my_bundle` that includes:

```yaml
sync:
  paths:
    - ../common
    - .
```

Then upon synchronization, the workspace will look like:
```
.
├── common
│   └── lib.py
└── my_bundle
    ├── databricks.yml
    └── notebook.py
```

If not set behavior remains identical.