Share test initializer in common helper function
These tests inadvertently were re-running mutators; the first time through
`loadTarget` and the second time by running `phases.Initialize()` themselves.
Some of the mutators that executed in `phases.Initializes()` are also run as
part of `loadTarget`. This is overdue a refactor to make it unambiguous what
runs when. Until then, this removes the duplicated execution.