Expand library globs relative to the sync root
Library glob expansion happens during deployment. Before that, all entries that
refer to local paths in resource definitions are made relative to the _sync
root_. Before #1694, they were made relative to the _bundle root_. This PR
didn't update the library glob expansion code to use the sync root path.

If you were using the sync paths setting with library globs, the CLI would fail
to expand the globs seeing as the code was using the wrong path to anchor those
globs on.

This change fixes the issue.