Add "output" flag to the bundle sync command
We want to use 'bundle sync' in the vscode extension before running a file
as an ad-hoc job (or through the context api). Right now we use bundle deploy
in these cases, but deploying bundle resources is not always expected when you
just want to quickly run a file. Sync make more sense in these cases, but we
still want to have verbose output to see what's happening.